Update README.md
This fixes a discrepancy between the readme and the linked source and prevents a compilation error that is hard to understand for people who are new to phoenix. 

This is a short term fix. The L sigil has been soft-deprecated. It would be better to properly fix the code to use the H sigil  but unfortunately I lack the skills.